CFY-1896 added test to a new file
